240 ta profile answer questions createanswer (#252)
* added instructorprofilepage with needviewer and basic test

* created comparer, cleaned up profile aesthetics

* added tests for instructor comparer and needviewer

* created allocation card need card tests

* improved aesthetics of comparer and made it user-accessible

* improved aesthetics of StudentCard and instructorprofile comparer

* changed titles and names to be Links

* changed aesthetics of sectionCard

* added coverage option to the project

* aesthetics reform to both instructor and ta profile page

* created tabs and pages and coursestakepage for ta

* added mockMATH125, fixed comparers and aesthetics

* fixed tests

* slight fix to NeedCard Additional Comments

* fixed hover:blue in allocationCard

* single Question questionnaire working

* completed tests for questions and answers

* created coordinator questionnaire version

* applied role authentication

* POST create question mapping works

* DELETE mapping integrated and tests fixed

* ta profile integrated with fetch call

* made deletion or updates of questions cascade delete any student answers

* questionnaire submission now resets all submissions

* fixed backend tests to work.

* frontend tests fixed adapted to work with changes

* attempt to refactor

* completed integration of backend with frontend profileanswers

* cleaned up code and made tests work

* updated postman, authentication in sidebar, changed routings

+++ app/Spring-Microservices/profileservice/src/main/java/com/infinity/profileservice/dtos/AnswerDto.java
@@ -1,5 +1,5 @@
 package com.infinity.profileservice.dtos;
 
-public record AnswerDto(Integer id, String description) {
+public record AnswerDto(Long id, String description) {
     
 }
